# Optimize campaign articles to use deferred joins and structured data loading