--- a/app/presenters/courses_presenter.rb
+++ b/app/presenters/courses_presenter.rb
@@ -41,13 +41,78 @@ class CoursesPresenter
   def campaign_articles
     return tag_articles if @tag
 
-    articles = campaign.articles_courses.tracked
-                       .includes(article: :wiki)
-                       .includes(:course).where(courses: { private: false })
-                       .paginate(page: @page, per_page: 100)
-    # Sorting can be particularly slow for large numbers of articles.
-    articles = articles.order('articles.deleted', character_sum: :desc) unless too_many_articles?
-    articles
+    {
+      articles_courses: paginated_articles_courses,
+      courses: load_courses.index_by(&:id),
+      articles: fetched_articles
+    }
+  end
+
+  # Fetch course IDs associated with the campaign
+  def campaigns_courses_ids
+    @campaigns_courses_ids ||= CampaignsCourses.where(campaign_id: campaign.id).select(:course_id).pluck(:course_id) # rubocop:disable Layout/LineLength
+  end
+
+  # Load only needed course data, grouped by ID
+  def load_courses
+    @courses ||= Course.where(id: campaigns_courses_ids, private: false).select(:id, :slug)
+  end
+
+  # Build subquery for ranked articles with pagination and optional sorting
+  def ranked_articles_subquery
+    ArticlesCourses
+      .includes(:article)
+      .where(course_id: load_courses.map(&:id), tracked: true)
+      .select(:id)
+      .then { |q| too_many_articles? ? q : q.order('articles.deleted ASC, articles_courses.character_sum DESC') } # rubocop:disable Layout/LineLength
+      .limit(per_page)
+      .offset(offset)
+  end
+
+  # Create deferred join scope using ranked subquery for performance
+  def articles_courses_scope
+    @articles_courses_scope ||= ArticlesCourses
+                                .joins("INNER JOIN (#{ranked_articles_subquery.to_sql}) AS ranked_articles USING (id)") # rubocop:disable Layout/LineLength
+                                .select(:article_id, :course_id, :character_sum, :references_count, :view_count) # rubocop:disable Layout/LineLength
+  end
+
+  # Fetch and index articles by ID for efficient lookup
+  def fetched_articles
+    @fetched_articles ||= begin
+      article_ids = articles_courses_scope.map(&:article_id).uniq
+      Article
+        .includes(:wiki)
+        .where(id: article_ids)
+        .select(:id, :title, :deleted, :wiki_id, :namespace)
+        .index_by(&:id)
+    end
+  end
+
+  # Create paginated collection for articles_courses data
+  def paginated_articles_courses
+    WillPaginate::Collection.create(current_page, per_page, total_articles_count) do |pager|
+      pager.replace(articles_courses_scope)
+    end
+  end
+
+  # Count total articles for pagination metadata
+  def total_articles_count
+    ArticlesCourses.where(course_id: load_courses.map(&:id), tracked: true).count
+  end
+
+  # Number of items per page
+  def per_page
+    100
+  end
+
+  # Current page number with fallback to 1
+  def current_page
+    @page.to_i < 1 ? 1 : @page.to_i
+  end
+
+  # Calculate offset for pagination
+  def offset
+    (current_page - 1) * per_page
   end
 
   # If there are too many articles, rendering a page of them can take a very long time.