## References

- [Deferred joins (Planetscale)](https://planetscale.com/learn/courses/mysql-for-developers/examples/deferred-joins)
- [Efficient MySQL pagination using deferred joins -  Aaron](https://aaronfrancis.com/2022/efficient-mysql-pagination-using-deferred-joins-15d0de14)
- [Deferred Join: A Deep Dive](https://hackmysql.com/deferred-join-deep-dive/)


## What this PR does


Refactors the campaign articles functionality to use deferred joins and structured data loading, significantly improving performance for large datasets while maintaining the same user experience and improving page load time.

The previous implementation had performance issues when dealing with large numbers of articles, particularly during pagination and sorting operations. The query approach was causing slow page loads for campaigns with extensive article collections.

## Solution
Implemented a deferred join pattern with structured data separation:

### Key Changes

#### 1. **Deferred Join Pattern**
- Replaced direct query with `ranked_articles_subquery` that pre-filters and ranks articles
- Uses `INNER JOIN` with the ranked subquery for optimal performance
- Maintains conditional sorting logic based on article count

#### 2. **Structured Data Loading**
- `campaign_articles` now returns a hash with three separate collections:
  - `articles_courses`: Paginated articles_courses data
  - `courses`: Indexed courses by ID for efficient lookup
  - `articles`: Indexed articles by ID for efficient lookup

#### 3. **Custom Pagination**
- Implements manual pagination using `WillPaginate::Collection`
- Proper offset/limit handling with `total_articles_count`
- Maintains existing pagination interface in views

#### 4. **Optimized Data Fetching**
- Uses `.select()` to load only required fields
- Implements instance variable caching for expensive operations
- Separates concerns with dedicated methods for each data type

### Performance Improvements
- **Query Optimization**: Deferred join reduces query complexity for large datasets
- **Memory Efficiency**: Loads only necessary fields and data structures
- **Caching**: Instance variables prevent redundant database calls
- **Scalability**: Handles large article counts without performance degradation

### View Updates
Updated templates to work with the new data structure:
- `articles.html.haml`: Uses structured data from the hash
- `_article_row.html.haml`: Accepts pre-loaded course and article objects
- Maintains existing UI/UX with no user-facing changes

## Technical Details

### Method Organization
The refactored presenter follows a logical data flow:
1. **Main interface**: `campaign_articles`
2. **Data sources**: `campaigns_courses_ids`, `load_courses`
3. **Query builders**: `ranked_articles_subquery`, `articles_courses_scope`
4. **Data fetchers**: `fetched_articles`
5. **Pagination**: `paginated_articles_courses`, `total_articles_count`
6. **Helpers**: `per_page`, `current_page`, `offset`

### Database Impact
- Reduces complex joins through deferred pattern
- Maintains existing indexes and relationships
- No schema changes required

## Testing
- Verify pagination works correctly for large datasets
- Confirm sorting behavior matches previous implementation
- Test performance with campaigns containing 1000+ articles
- Ensure view rendering remains consistent


## Screenshots  `(WITH SORTING)`

**Before:**

**`PAGE 1`**  **(4841 ms)**

![Screenshot from 2025-07-08 22-25-49](https://github.com/user-attachments/assets/d23b0996-18bc-4bf8-9c6f-a15a2959daab)

**`PAGE 1197`**  **(5870 ms)**

![Screenshot from 2025-07-08 22-27-28](https://github.com/user-attachments/assets/021b5cda-7630-4cbe-b5af-25406d2910a1)


**After:** 

**`PAGE 1`**  **(732 ms)**

![Screenshot from 2025-07-08 22-20-37](https://github.com/user-attachments/assets/6dfdb91b-ff24-410f-87b1-c29885f7c5a7)


**`PAGE 1197`**  **(831 ms)**

![Screenshot from 2025-07-08 22-22-26](https://github.com/user-attachments/assets/bcd53c29-c9ca-4037-b3d1-8e81b5fec3ed)


## Screenshots  `(WITHOUT SORTING)`


**BEFORE**

**`PAGE 1`**  **(430 ms)**


![Screenshot from 2025-07-09 17-52-56](https://github.com/user-attachments/assets/f262a555-515b-417c-adc9-42621cc56e1b)



**`PAGE 1197`** **(2068 ms)**

![Screenshot from 2025-07-09 17-50-41](https://github.com/user-attachments/assets/02c71022-88dd-4227-8581-0a89ab0ddaf6)



**AFTER**


**`PAGE 1`**  **(307 ms)**

![Screenshot from 2025-07-09 17-26-46](https://github.com/user-attachments/assets/60071fc7-3232-4cbc-95a2-111c57d7b148)


**`PAGE 1197`**  **(268 ms)**

![Screenshot from 2025-07-09 17-28-31](https://github.com/user-attachments/assets/7cda7c3e-01a5-4132-aa34-4ab79ffd19e2)