@Abishekcs 2025-07-09
> This significantly increases the size and complexity of CoursesPresenter. Is there a clean way to extract the code for the main complex query to a separate file?

I can try to do that .