@ragesoss 2025-07-11
This looks good to me. It still adds a lot of code to CoursesPresenter, which makes it larger and more complex than I'd like... especially because it is used in several different contexts but the new code is only for use with campaign views. Most of the logic was shared between the different uses before, but this change should prompt us to think about how to better organize the code so that we don't have one file pulled in multiple directions, perhaps by extracting the shared code between the use cases and implementing a separate CampaignCoursesPresenter that includes the campaign-specific parts. However, that's not necessary immediately, and I'm fine with merging in the current state.

@Formasitchijoh anything you spot that should be improved before merging?